started refactoring code and optimising SQL queries
