+++ app/routers/developer.py
@@ -4,6 +4,7 @@ from app.pymysql.databaseConnection import get_db_connection
 from typing import Annotated
 from app.dependencies import get_current_user
 from app.models.User import User
+from app.utils.db_utils import get_info_list
 
 router = APIRouter()
 
@@ -18,26 +19,8 @@ class Developer(BaseModel):
 # get all developers
 @router.get("/developers/")
 def get_developers():
-    try:
-        # make a database connection
-        connection = get_db_connection()
-        # create a cursor object
-        cursor = connection.cursor()
-        cursor.execute("SELECT * FROM developer")
-        rows = cursor.fetchall()
-    except Exception as e:
-        print(e)
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail={"success": False, "message": "An error occurred"},
-        )
-    finally:
-        connection.close()
-
-    # on successful operation, send status 200 and messages
-    raise HTTPException(
-        status_code=status.HTTP_200_OK, detail={"success": True, "rows": rows}
-    )
+    query = "SELECT developer_id, name FROM developer"
+    get_info_list(query)
 
 
 # fetch all data about a single developer
